make sure input array is float

+++ dst.py
@@ -170,9 +170,9 @@ for pol, comps in zip([False, True], ["T", "QU"]):
                 'U' : np.zeros(len(data['U']), dtype=np.double),
                 }
                 try:
-                    signalremove(signal_removed['Q'], signal_removed['U'], data['Q'], data['U'], tmap_local.array, umap_local.array, data['q_channel_w']['Q'], data['q_channel_w']['U'], data['u_channel_w']['Q'], data['u_channel_w']['U'], pix)
+                    signalremove(signal_removed['Q'], signal_removed['U'], data['Q'].astype(np.float32), data['U'].astype(np.float32), tmap_local.array, umap_local.array, data['q_channel_w']['Q'], data['q_channel_w']['U'], data['u_channel_w']['Q'], data['u_channel_w']['U'], pix)
                 except exceptions.ValueError: # byteswap in case input h5 has wrong endianness
-                    signalremove(signal_removed['Q'], signal_removed['U'], data['Q'].byteswap().newbyteorder(), data['U'].byteswap().newbyteorder(), tmap_local.array, umap_local.array, data['q_channel_w']['Q'], data['q_channel_w']['U'], data['u_channel_w']['Q'], data['u_channel_w']['U'], pix)
+                    signalremove(signal_removed['Q'], signal_removed['U'], data['Q'].byteswap().newbyteorder().astype(np.float32), data['U'].byteswap().newbyteorder().astype(np.float32), tmap_local.array, umap_local.array, data['q_channel_w']['Q'], data['q_channel_w']['U'], data['u_channel_w']['Q'], data['u_channel_w']['U'], pix)
 
                 assert len(baseline_lengths) == len(RHS.array[0][:num_baselines])
                 accumulate(signal_removed['Q'], baseline_lengths, RHS.array[0][:num_baselines])
@@ -188,7 +188,7 @@ for pol, comps in zip([False, True], ["T", "QU"]):
                 l.info("Remove signal")
                 signal_removed = np.zeros(len(data['T']), dtype=np.double)
                 try:
-                    signalremovet(signal_removed, data['T'], tmap_local.array, pix)
+                    signalremovet(signal_removed, data['T'].astype(np.float32), tmap_local.array, pix)
                 except exceptions.ValueError: # byteswap in case input h5 has wrong endianness
                     signalremovet(signal_removed, data['T'].byteswap().newbyteorder(), tmap_local.array, pix)
